Add verify dif test for cw mac
About to fix its prg init when using both dpf and dcf

+++ src/dpf_test.cc
@@ -30,7 +30,7 @@ class DpfTest : public ::testing::Test {
   }
 
   static constexpr uint16_t kAlpha = 107;
-  static constexpr uint16_t kAlphaBitlen = 16;
+  static constexpr int kAlphaBitlen = 16;
   static constexpr __uint128_t kBeta = 604;
 
   uint8_t *kS0s;
